requirements.txt
Django==1.8
Markdown==2.6.2
Pillow==2.8.1
argparse==1.2.1
django-filter==0.9.2
django-filters==0.1.0
djangorestframework==3.1.1
wsgiref==0.1.2
